refactor(frontend): rename api/loki.ts to api/query-response.ts
NETOBSERV-1828 The file web/src/api/loki.ts contained generic response types
(AggregatedQueryResponse, Stats, TopologyMetrics, NetflowMetrics, etc.)
shared by both Loki and Prometheus query paths. Having it named after
Loki was misleading since these types have nothing Loki-specific about
them.

Rename to query-response.ts and update all ~40 import sites across
components, utils, models, and test files.
