+++ web/src/components/query-summary/metrics-query-summary.tsx
@@ -1,6 +1,6 @@
 import { Card } from '@patternfly/react-core';
 import * as React from 'react';
-import { NetflowMetrics, Stats } from '../../api/loki';
+import { NetflowMetrics, Stats } from '../../api/query-response';
 import { Warning } from '../../model/warnings';
 import { MetricsQuerySummaryContent } from './metrics-query-summary-content';
 import './query-summary.css';
