+++ web/src/components/query-summary/summary-panel.tsx
@@ -11,7 +11,7 @@ import {
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Record } from '../../api/ipfix';
-import { NetflowMetrics, Stats } from '../../api/loki';
+import { NetflowMetrics, Stats } from '../../api/query-response';
 import { RecordType } from '../../model/flow-query';
 import { Warning } from '../../model/warnings';
 import { TimeRange } from '../../utils/datetime';
